refactor: split model catalog by provider family
