+++ src/model_catalog.rs
@@ -1386,591 +1386,14 @@ fn reasoning_effort_options(
     options.iter().map(|option| (*option).to_string()).collect()
 }
 
-fn catalog_model(
-    provider: &str,
-    model: &str,
-    display_name: &str,
-    context_window_tokens: usize,
-    max_output_tokens: u32,
-    supports_reasoning: bool,
-    image_input: bool,
-) -> BuiltInModelMetadata {
-    let model_ref = ModelRef::new(provider_id(provider), model);
-    BuiltInModelMetadata {
-        default_verbosity: default_verbosity_for_model(&model_ref),
-        model_ref,
-        display_name: display_name.into(),
-        description: format!(
-            "Holon built-in runtime metadata for the {provider}/{model} compatible provider model."
-        ),
-        context_window_tokens: Some(context_window_tokens),
-        effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-        auto_compact_token_limit: None,
-        default_max_output_tokens: Some(max_output_tokens),
-        max_output_tokens_upper_limit: Some(max_output_tokens),
-        tool_output_truncation_estimated_tokens: Some(
-            DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-        ),
-        capabilities: ModelCapabilityFlags {
-            image_input,
-            supports_reasoning,
-            ..ModelCapabilityFlags::default()
-        },
-        reasoning_effort_options: Vec::new(),
-        source: ModelMetadataSource::BuiltInCatalog,
-        endpoint: None,
-    }
-}
+mod providers;
+#[cfg(test)]
+use providers::common::catalog_model;
 
-fn opencode_go_model(
-    model: &str,
-    display_name: &str,
-    context_window_tokens: usize,
-    max_output_tokens: Option<u32>,
-    supports_reasoning: bool,
-    image_input: bool,
-    endpoint: Option<&str>,
-) -> BuiltInModelMetadata {
-    let model_ref = ModelRef::new(provider_id("opencode-go"), model);
-    BuiltInModelMetadata {
-        default_verbosity: default_verbosity_for_model(&model_ref),
-        model_ref,
-        display_name: display_name.into(),
-        description: format!(
-            "Holon conservative built-in metadata for the OpenCode Go {model} route."
-        ),
-        context_window_tokens: Some(context_window_tokens),
-        effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-        auto_compact_token_limit: None,
-        default_max_output_tokens: max_output_tokens,
-        max_output_tokens_upper_limit: max_output_tokens,
-        tool_output_truncation_estimated_tokens: Some(
-            DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-        ),
-        capabilities: ModelCapabilityFlags {
-            image_input,
-            supports_reasoning,
-            ..ModelCapabilityFlags::default()
-        },
-        reasoning_effort_options: Vec::new(),
-        source: ModelMetadataSource::ConservativeBuiltin,
-        endpoint: endpoint.map(|endpoint| {
-            ProviderEndpointId::parse(endpoint).expect("valid OpenCode Go endpoint id")
-        }),
-    }
-}
+pub(crate) use providers::is_tencent_tokenhub_model_id;
 
-fn venice_model(
-    model: &str,
-    display_name: &str,
-    context_window_tokens: usize,
-    max_output_tokens: u32,
-    supports_reasoning: bool,
-    image_input: bool,
-) -> BuiltInModelMetadata {
-    let model_ref = ModelRef::new(provider_id("venice"), model);
-    let capabilities = ModelCapabilityFlags {
-        image_input,
-        supports_reasoning,
-        ..ModelCapabilityFlags::default()
-    };
-    BuiltInModelMetadata {
-        default_verbosity: default_verbosity_for_model(&model_ref),
-        model_ref: model_ref.clone(),
-        display_name: display_name.into(),
-        description: format!(
-            "Holon conservative built-in metadata for the Venice {model} text model."
-        ),
-        context_window_tokens: Some(context_window_tokens),
-        effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-        auto_compact_token_limit: None,
-        default_max_output_tokens: Some(max_output_tokens),
-        max_output_tokens_upper_limit: Some(max_output_tokens),
-        tool_output_truncation_estimated_tokens: Some(
-            DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-        ),
-        reasoning_effort_options: reasoning_effort_options(&model_ref, None, &capabilities),
-        capabilities,
-        source: ModelMetadataSource::ConservativeBuiltin,
-        endpoint: None,
-    }
-}
-
-fn chutes_model_without_published_capabilities(
-    model: &str,
-    context_window_tokens: usize,
-) -> BuiltInModelMetadata {
-    let model_ref = ModelRef::new(provider_id("chutes"), model);
-    BuiltInModelMetadata {
-        default_verbosity: default_verbosity_for_model(&model_ref),
-        model_ref,
-        display_name: model.into(),
-        description: format!(
-            "Holon conservative built-in metadata for the Chutes {model} model; the public model entry does not publish complete capability fields."
-        ),
-        context_window_tokens: Some(context_window_tokens),
-        effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-        auto_compact_token_limit: None,
-        default_max_output_tokens: None,
-        max_output_tokens_upper_limit: None,
-        tool_output_truncation_estimated_tokens: Some(
-            DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-        ),
-        capabilities: ModelCapabilityFlags::default(),
-        reasoning_effort_options: Vec::new(),
-        source: ModelMetadataSource::ConservativeBuiltin,
-        endpoint: None,
-    }
-}
-
-fn arcee_model(model: &str, display_name: &str) -> BuiltInModelMetadata {
-    let model_ref = ModelRef::new(provider_id("arcee"), model);
-    BuiltInModelMetadata {
-        default_verbosity: default_verbosity_for_model(&model_ref),
-        model_ref,
-        display_name: display_name.into(),
-        description: format!(
-            "Holon conservative built-in metadata for the Arcee hosted {model} model."
-        ),
-        context_window_tokens: Some(131_072),
-        effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-        auto_compact_token_limit: None,
-        default_max_output_tokens: None,
-        max_output_tokens_upper_limit: None,
-        tool_output_truncation_estimated_tokens: Some(
-            DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-        ),
-        capabilities: ModelCapabilityFlags::default(),
-        reasoning_effort_options: Vec::new(),
-        source: ModelMetadataSource::ConservativeBuiltin,
-        endpoint: None,
-    }
-}
-
-fn huggingface_gpt_oss_model() -> BuiltInModelMetadata {
-    let model_ref = ModelRef::new(provider_id("huggingface"), "openai/gpt-oss-120b");
-    BuiltInModelMetadata {
-        default_verbosity: default_verbosity_for_model(&model_ref),
-        model_ref,
-        display_name: "OpenAI GPT OSS 120B".into(),
-        description:
-            "Holon conservative built-in metadata for the Hugging Face Inference Providers quick-start chat model."
-                .into(),
-        context_window_tokens: Some(131_072),
-        effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-        auto_compact_token_limit: None,
-        default_max_output_tokens: None,
-        max_output_tokens_upper_limit: None,
-        tool_output_truncation_estimated_tokens: Some(
-            DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-        ),
-        capabilities: ModelCapabilityFlags {
-            supports_reasoning: true,
-            ..ModelCapabilityFlags::default()
-        },
-        reasoning_effort_options: vec!["low".into(), "medium".into(), "high".into()],
-        source: ModelMetadataSource::ConservativeBuiltin,
-        endpoint: None,
-    }
-}
-
-fn kilocode_auto_model(
-    model: &str,
-    display_name: &str,
-    context_window_tokens: usize,
-    max_output_tokens: u32,
-    image_input: bool,
-) -> BuiltInModelMetadata {
-    let model_ref = ModelRef::new(provider_id("kilocode"), model);
-    BuiltInModelMetadata {
-        default_verbosity: default_verbosity_for_model(&model_ref),
-        model_ref,
-        display_name: display_name.into(),
-        description: format!(
-            "Holon conservative built-in metadata for the Kilo Gateway {display_name} virtual model."
-        ),
-        context_window_tokens: Some(context_window_tokens),
-        effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-        auto_compact_token_limit: None,
-        default_max_output_tokens: None,
-        max_output_tokens_upper_limit: Some(max_output_tokens),
-        tool_output_truncation_estimated_tokens: Some(
-            DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-        ),
-        capabilities: ModelCapabilityFlags {
-            image_input,
-            supports_reasoning: true,
-            ..ModelCapabilityFlags::default()
-        },
-        reasoning_effort_options: Vec::new(),
-        source: ModelMetadataSource::ConservativeBuiltin,
-        endpoint: None,
-    }
-}
-
-fn fireworks_model(
-    model: &str,
-    display_name: &str,
-    context_window_tokens: Option<usize>,
-    supports_reasoning: bool,
-    image_input: bool,
-) -> BuiltInModelMetadata {
-    let model_ref = ModelRef::new(provider_id("fireworks"), model);
-    BuiltInModelMetadata {
-        default_verbosity: default_verbosity_for_model(&model_ref),
-        model_ref,
-        display_name: display_name.into(),
-        description: format!(
-            "Holon conservative built-in metadata for the Fireworks serverless {model} model."
-        ),
-        context_window_tokens,
-        effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-        auto_compact_token_limit: None,
-        default_max_output_tokens: None,
-        max_output_tokens_upper_limit: None,
-        tool_output_truncation_estimated_tokens: Some(
-            DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-        ),
-        capabilities: ModelCapabilityFlags {
-            image_input,
-            supports_reasoning,
-            ..ModelCapabilityFlags::default()
-        },
-        reasoning_effort_options: Vec::new(),
-        source: ModelMetadataSource::ConservativeBuiltin,
-        endpoint: None,
-    }
-}
-
-fn nvidia_model(
-    model: &str,
-    display_name: &str,
-    context_window_tokens: usize,
-    supports_reasoning: bool,
-    image_input: bool,
-) -> BuiltInModelMetadata {
-    let model_ref = ModelRef::new(provider_id("nvidia"), model);
-    BuiltInModelMetadata {
-        default_verbosity: default_verbosity_for_model(&model_ref),
-        model_ref,
-        display_name: display_name.into(),
-        description: format!(
-            "Holon conservative built-in metadata for the NVIDIA-hosted {model} model."
-        ),
-        context_window_tokens: Some(context_window_tokens),
-        effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-        auto_compact_token_limit: None,
-        default_max_output_tokens: None,
-        max_output_tokens_upper_limit: None,
-        tool_output_truncation_estimated_tokens: Some(
-            DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-        ),
-        capabilities: ModelCapabilityFlags {
-            image_input,
-            supports_reasoning,
-            ..ModelCapabilityFlags::default()
-        },
-        reasoning_effort_options: Vec::new(),
-        source: ModelMetadataSource::ConservativeBuiltin,
-        endpoint: None,
-    }
-}
-
-fn together_model(
-    model: &str,
-    display_name: &str,
-    context_window_tokens: usize,
-    supports_reasoning: bool,
-    image_input: bool,
-) -> BuiltInModelMetadata {
-    let model_ref = ModelRef::new(provider_id("together"), model);
-    BuiltInModelMetadata {
-        default_verbosity: default_verbosity_for_model(&model_ref),
-        model_ref,
-        display_name: display_name.into(),
-        description: format!(
-            "Holon conservative built-in metadata for the Together serverless {model} model."
-        ),
-        context_window_tokens: Some(context_window_tokens),
-        effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-        auto_compact_token_limit: None,
-        default_max_output_tokens: None,
-        max_output_tokens_upper_limit: None,
-        tool_output_truncation_estimated_tokens: Some(
-            DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-        ),
-        capabilities: ModelCapabilityFlags {
-            image_input,
-            supports_reasoning,
-            ..ModelCapabilityFlags::default()
-        },
-        reasoning_effort_options: Vec::new(),
-        source: ModelMetadataSource::ConservativeBuiltin,
-        endpoint: None,
-    }
-}
-
-fn stepfun_model(
-    provider: &str,
-    model: &str,
-    display_name: &str,
-    image_input: bool,
-) -> BuiltInModelMetadata {
-    let model_ref = ModelRef::new(provider_id(provider), model);
-    BuiltInModelMetadata {
-        default_verbosity: default_verbosity_for_model(&model_ref),
-        model_ref,
-        display_name: display_name.into(),
-        description: format!("Holon built-in runtime metadata for the StepFun {model} model."),
-        context_window_tokens: Some(262_144),
-        effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-        auto_compact_token_limit: None,
-        default_max_output_tokens: None,
-        max_output_tokens_upper_limit: None,
-        tool_output_truncation_estimated_tokens: Some(
-            DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-        ),
-        capabilities: ModelCapabilityFlags {
-            image_input,
-            supports_reasoning: true,
-            ..ModelCapabilityFlags::default()
-        },
-        reasoning_effort_options: Vec::new(),
-        source: ModelMetadataSource::BuiltInCatalog,
-        endpoint: None,
-    }
-}
-
-fn tencent_tokenhub_model(
-    provider: &str,
-    model: &str,
-    display_name: &str,
-    context_window_tokens: Option<usize>,
-    max_output_tokens: Option<u32>,
-    supports_reasoning: bool,
-    image_input: bool,
-) -> BuiltInModelMetadata {
-    let model_ref = ModelRef::new(provider_id(provider), model);
-    BuiltInModelMetadata {
-        default_verbosity: default_verbosity_for_model(&model_ref),
-        model_ref,
-        display_name: display_name.into(),
-        description: format!(
-            "Holon conservative built-in metadata for the Tencent TokenHub {model} model."
-        ),
-        context_window_tokens,
-        effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-        auto_compact_token_limit: None,
-        default_max_output_tokens: max_output_tokens,
-        max_output_tokens_upper_limit: max_output_tokens,
-        tool_output_truncation_estimated_tokens: Some(
-            DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-        ),
-        capabilities: ModelCapabilityFlags {
-            image_input,
-            supports_reasoning,
-            ..ModelCapabilityFlags::default()
-        },
-        reasoning_effort_options: Vec::new(),
-        source: ModelMetadataSource::ConservativeBuiltin,
-        endpoint: None,
-    }
-}
-
-type TencentTokenHubModelSpec = (
-    &'static str,
-    &'static str,
-    Option<usize>,
-    Option<u32>,
-    bool,
-    bool,
-);
-
-const TENCENT_TOKENHUB_MODELS: &[TencentTokenHubModelSpec] = &[
-    ("hy3", "Hy3", Some(256_000), Some(128_000), true, false),
-    (
-        "hy3-preview",
-        "Hy3 Preview",
-        Some(256_000),
-        Some(128_000),
-        true,
-        false,
-    ),
-    ("hy-mt2-pro", "Hy-MT2-Pro", None, None, false, false),
-    ("hy-mt2-plus", "Hy-MT2-Plus", None, None, false, false),
-    ("hy-mt2-lite", "Hy-MT2-Lite", None, None, false, false),
-    (
-        "hunyuan-role-latest",
-        "Hy-Role-Latest",
-        None,
-        None,
-        false,
-        false,
-    ),
-    ("hy-role", "Hy-Role", None, None, false, false),
-    (
-        "deepseek-v4-flash",
-        "DeepSeek V4 Flash",
-        Some(1_000_000),
-        Some(384_000),
-        true,
-        false,
-    ),
-    (
-        "deepseek-v4-pro",
-        "DeepSeek V4 Pro",
-        Some(1_000_000),
-        Some(384_000),
-        true,
-        false,
-    ),
-    (
-        "deepseek-v3.2",
-        "DeepSeek V3.2",
-        Some(128_000),
-        Some(32_768),
-        true,
-        false,
-    ),
-    (
-        "glm-5.2",
-        "GLM-5.2",
-        Some(1_000_000),
-        Some(131_072),
-        true,
-        false,
-    ),
-    (
-        "glm-5.1",
-        "GLM-5.1",
-        Some(202_800),
-        Some(131_072),
-        true,
-        false,
-    ),
-    (
-        "glm-5v-turbo",
-        "GLM-5V Turbo",
-        Some(202_800),
-        Some(131_072),
-        true,
-        true,
-    ),
-    (
-        "glm-5-turbo",
-        "GLM-5 Turbo",
-        Some(202_800),
-        Some(131_072),
-        true,
-        false,
-    ),
-    ("glm-5", "GLM-5", Some(202_800), Some(131_072), true, false),
-    (
-        "kimi-k2.7-code-highspeed",
-        "Kimi K2.7 Code HighSpeed",
-        Some(262_144),
-        Some(262_144),
-        true,
-        true,
-    ),
-    (
-        "kimi-k2.7-code",
-        "Kimi K2.7 Code",
-        Some(262_144),
-        Some(262_144),
-        true,
-        true,
-    ),
-    (
-        "kimi-k2.6",
-        "Kimi K2.6",
-        Some(262_144),
-        Some(262_144),
-        true,
-        true,
-    ),
-    (
-        "kimi-k2.5",
-        "Kimi K2.5",
-        Some(262_144),
-        Some(262_144),
-        true,
-        true,
-    ),
-    (
-        "minimax-m3",
-        "MiniMax M3",
-        Some(1_000_000),
-        None,
-        true,
-        true,
-    ),
-    (
-        "minimax-m2.7",
-        "MiniMax M2.7",
-        Some(204_800),
-        None,
-        true,
-        false,
-    ),
-    (
-        "minimax-m2.5",
-        "MiniMax M2.5",
-        Some(196_608),
-        Some(32_768),
-        true,
-        false,
-    ),
-    (
-        "qwen3.5-flash",
-        "Qwen3.5 Flash",
-        Some(1_000_000),
-        Some(65_536),
-        true,
-        true,
-    ),
-    (
-        "qwen3.5-plus",
-        "Qwen3.5 Plus",
-        Some(1_000_000),
-        Some(65_536),
-        true,
-        true,
-    ),
-    ("youtu-vita", "YT-VITA", None, None, false, true),
-    (
-        "hy-vision-2.0-instruct",
-        "HY-Vision-2.0-Instruct",
-        None,
-        None,
-        false,
-        true,
-    ),
-    (
-        "hunyuan-t1-vision-20250916",
-        "HY-Vision-1.5-Thinking",
-        None,
-        None,
-        true,
-        true,
-    ),
-];
-
-pub(crate) fn is_tencent_tokenhub_model_id(id: &str) -> bool {
-    TENCENT_TOKENHUB_MODELS.iter().any(|model| model.0 == id)
-}
-
-fn tencent_tokenhub_model_entries() -> Vec<BuiltInModelMetadata> {
-    ["tencent-tokenhub", "tencent-tokenhub-messages"]
-        .into_iter()
-        .flat_map(|provider| {
-            TENCENT_TOKENHUB_MODELS.iter().map(move |model| {
-                tencent_tokenhub_model(
-                    provider, model.0, model.1, model.2, model.3, model.4, model.5,
-                )
-            })
-        })
-        .collect()
+fn built_in_entries() -> Vec<BuiltInModelMetadata> {
+    providers::built_in_entries()
 }
 
 fn is_turn_default_candidate(entry: &BuiltInModelMetadata) -> bool {
@@ -1981,2262 +1404,8 @@ fn is_turn_default_candidate(entry: &BuiltInModelMetadata) -> bool {
         || entry.capabilities.interactive_exec
 }
 
-fn built_in_entries() -> Vec<BuiltInModelMetadata> {
-    let mut entries = vec![
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::anthropic(), "claude-fable-5"),
-            display_name: "Claude Fable 5".into(),
-            description: "Anthropic Claude Fable 5 runtime defaults aligned with the official model overview.".into(),
-            context_window_tokens: Some(1_000_000),
-            effective_context_window_percent: 90,
-            auto_compact_token_limit: Some(900_000),
-            default_max_output_tokens: Some(128_000),
-            max_output_tokens_upper_limit: Some(128_000),
-            default_verbosity: None,
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::anthropic(), "claude-opus-4-8"),
-            display_name: "Claude Opus 4.8".into(),
-            description: "Anthropic Claude Opus 4.8 runtime defaults aligned with the official model overview.".into(),
-            context_window_tokens: Some(1_000_000),
-            effective_context_window_percent: 90,
-            auto_compact_token_limit: Some(900_000),
-            default_max_output_tokens: Some(128_000),
-            max_output_tokens_upper_limit: Some(128_000),
-            default_verbosity: None,
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::anthropic(), "claude-sonnet-5"),
-            display_name: "Claude Sonnet 5".into(),
-            description: "Anthropic Claude Sonnet 5 runtime defaults aligned with the official model overview.".into(),
-            context_window_tokens: Some(1_000_000),
-            effective_context_window_percent: 90,
-            auto_compact_token_limit: Some(900_000),
-            default_max_output_tokens: Some(128_000),
-            max_output_tokens_upper_limit: Some(128_000),
-            default_verbosity: None,
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::anthropic(), "claude-haiku-4-5"),
-            display_name: "Claude Haiku 4.5".into(),
-            description: "Anthropic Claude Haiku 4.5 runtime defaults aligned with the official model overview.".into(),
-            context_window_tokens: Some(200_000),
-            effective_context_window_percent: 90,
-            auto_compact_token_limit: Some(180_000),
-            default_max_output_tokens: Some(64_000),
-            max_output_tokens_upper_limit: Some(64_000),
-            default_verbosity: None,
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::openai_codex(), "gpt-5.5"),
-            display_name: "GPT-5.5 (Codex)".into(),
-            description: "Codex runtime defaults mirrored from the local OpenAI Codex model metadata contract.".into(),
-            context_window_tokens: Some(272_000),
-            effective_context_window_percent: 95,
-            auto_compact_token_limit: None,
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: Some(ModelVerbosity::Low),
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                image_generation: true,
-                interactive_exec: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::openai_codex(), "gpt-5.6-sol"),
-            display_name: "GPT-5.6-Sol (Codex)".into(),
-            description: "Codex runtime defaults mirrored from the local OpenAI Codex model metadata contract.".into(),
-            context_window_tokens: Some(372_000),
-            effective_context_window_percent: 95,
-            auto_compact_token_limit: None,
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: Some(ModelVerbosity::Low),
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                image_generation: true,
-                interactive_exec: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::openai_codex(), "gpt-5.6-terra"),
-            display_name: "GPT-5.6-Terra (Codex)".into(),
-            description: "Codex runtime defaults mirrored from the local OpenAI Codex model metadata contract.".into(),
-            context_window_tokens: Some(372_000),
-            effective_context_window_percent: 95,
-            auto_compact_token_limit: None,
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: Some(ModelVerbosity::Low),
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                image_generation: true,
-                interactive_exec: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::openai_codex(), "gpt-5.6-luna"),
-            display_name: "GPT-5.6-Luna (Codex)".into(),
-            description: "Codex runtime defaults mirrored from the local OpenAI Codex model metadata contract.".into(),
-            context_window_tokens: Some(372_000),
-            effective_context_window_percent: 95,
-            auto_compact_token_limit: None,
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: Some(ModelVerbosity::Low),
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                image_generation: true,
-                interactive_exec: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::openai_codex(), "gpt-5.4"),
-            display_name: "GPT-5.4 (Codex)".into(),
-            description: "Codex runtime defaults mirrored from the local OpenAI Codex model metadata contract.".into(),
-            context_window_tokens: Some(272_000),
-            effective_context_window_percent: 95,
-            auto_compact_token_limit: None,
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: Some(ModelVerbosity::Low),
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                image_generation: true,
-                interactive_exec: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::openai_codex(), "gpt-5.4-mini"),
-            display_name: "GPT-5.4-Mini (Codex)".into(),
-            description: "Codex runtime defaults mirrored from the local OpenAI Codex model metadata contract.".into(),
-            context_window_tokens: Some(272_000),
-            effective_context_window_percent: 95,
-            auto_compact_token_limit: None,
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: Some(ModelVerbosity::Medium),
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                image_generation: true,
-                interactive_exec: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::openai_codex(), "gpt-5.3-codex-spark"),
-            display_name: "GPT-5.3 Codex Spark (Codex)".into(),
-            description: "Codex Spark runtime defaults mirrored from the local OpenAI Codex model metadata contract.".into(),
-            context_window_tokens: Some(128_000),
-            effective_context_window_percent: 95,
-            auto_compact_token_limit: None,
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: Some(ModelVerbosity::Low),
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                interactive_exec: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::openai(), "gpt-image-2"),
-            display_name: "GPT Image 2".into(),
-            description: "OpenAI image generation model for the Images API.".into(),
-            context_window_tokens: None,
-            effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-            auto_compact_token_limit: None,
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: None,
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_generation: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::openai(), "gpt-5.4"),
-            display_name: "GPT-5.4".into(),
-            description: "Conservative GPT-5.4 runtime defaults aligned with local Codex model behavior.".into(),
-            context_window_tokens: Some(272_000),
-            effective_context_window_percent: 95,
-            auto_compact_token_limit: None,
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: None,
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::ConservativeBuiltin,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::openai(), "gpt-5.3"),
-            display_name: "GPT-5.3".into(),
-            description: "Conservative GPT-5.3 runtime defaults used when explicit model metadata is not available locally.".into(),
-            context_window_tokens: Some(128_000),
-            effective_context_window_percent: 95,
-            auto_compact_token_limit: None,
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: None,
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::ConservativeBuiltin,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(ProviderId::openai(), "gpt-5.4-mini"),
-            display_name: "GPT-5.4 Mini".into(),
-            description: "Conservative GPT-5.4 Mini runtime defaults used when explicit model metadata is not available locally.".into(),
-            context_window_tokens: Some(128_000),
-            effective_context_window_percent: 95,
-            auto_compact_token_limit: None,
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: None,
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::ConservativeBuiltin,
-            endpoint: None,
-        },
-    ];
-    entries.extend(compatible_provider_model_entries());
-    entries
-}
-
 fn default_verbosity_for_model(model_ref: &ModelRef) -> Option<ModelVerbosity> {
     (model_ref.provider == ProviderId::openai_codex()).then_some(ModelVerbosity::Low)
-}
-
-fn compatible_provider_model_entries() -> Vec<BuiltInModelMetadata> {
-    let mut entries = vec![
-        catalog_model(
-            "openai",
-            "gpt-5.6-sol",
-            "GPT-5.6 Sol",
-            372_000,
-            128_000,
-            true,
-            true,
-        ),
-        catalog_model(
-            "openai",
-            "gpt-5.6-terra",
-            "GPT-5.6 Terra",
-            372_000,
-            128_000,
-            true,
-            true,
-        ),
-        catalog_model(
-            "openai",
-            "gpt-5.6-luna",
-            "GPT-5.6 Luna",
-            372_000,
-            128_000,
-            true,
-            true,
-        ),
-        catalog_model(
-            "gemini",
-            "gemini-3.5-flash",
-            "Gemini 3.5 Flash",
-            1_048_576,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "gemini",
-            "gemini-3.1-pro-preview",
-            "Gemini 3.1 Pro Preview",
-            1_048_576,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "gemini",
-            "gemini-3.1-flash-lite",
-            "Gemini 3.1 Flash-Lite",
-            1_048_576,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "gemini",
-            "gemini-2.5-pro",
-            "Gemini 2.5 Pro",
-            1_048_576,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "gemini",
-            "gemini-2.5-flash",
-            "Gemini 2.5 Flash",
-            1_048_576,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "gemini",
-            "gemini-2.5-flash-lite",
-            "Gemini 2.5 Flash-Lite",
-            1_048_576,
-            65_536,
-            true,
-            true,
-        ),
-        arcee_model("trinity-mini", "Trinity Mini 26B"),
-        arcee_model("trinity-large-preview", "Trinity Large Preview"),
-        catalog_model(
-            "chutes",
-            "moonshotai/Kimi-K2.6-TEE",
-            "moonshotai/Kimi-K2.6-TEE",
-            262_144,
-            65_535,
-            true,
-            true,
-        ),
-        catalog_model("chutes", "zai-org/GLM-5.2-TEE", "zai-org/GLM-5.2-TEE", 1_048_576, 65_535, true, false),
-        catalog_model("chutes", "zai-org/GLM-5.1-TEE", "zai-org/GLM-5.1-TEE", 202_752, 65_535, true, false),
-        catalog_model("chutes", "zai-org/GLM-5-TEE", "zai-org/GLM-5-TEE", 202_752, 65_535, true, false),
-        catalog_model("chutes", "deepseek-ai/DeepSeek-V3.2-TEE", "deepseek-ai/DeepSeek-V3.2-TEE", 131_072, 65_536, true, false),
-        catalog_model("chutes", "MiniMaxAI/MiniMax-M2.5-TEE", "MiniMaxAI/MiniMax-M2.5-TEE", 196_608, 65_536, true, false),
-        catalog_model("chutes", "moonshotai/Kimi-K2.5-TEE", "moonshotai/Kimi-K2.5-TEE", 262_144, 65_535, true, true),
-        catalog_model("chutes", "Qwen/Qwen3.6-27B-TEE", "Qwen/Qwen3.6-27B-TEE", 262_144, 65_536, true, true),
-        catalog_model("chutes", "Qwen/Qwen3.5-397B-A17B-TEE", "Qwen/Qwen3.5-397B-A17B-TEE", 262_144, 65_536, true, true),
-        catalog_model("chutes", "Qwen/Qwen3-235B-A22B-Thinking-2507-TEE", "Qwen/Qwen3-235B-A22B-Thinking-2507-TEE", 262_144, 262_144, true, false),
-        catalog_model("chutes", "Qwen/Qwen3-32B-TEE", "Qwen/Qwen3-32B-TEE", 40_960, 40_960, true, false),
-        catalog_model("chutes", "google/gemma-4-31B-turbo-TEE", "google/gemma-4-31B-turbo-TEE", 131_072, 65_536, true, true),
-        chutes_model_without_published_capabilities(
-            "unsloth/Mistral-Nemo-Instruct-2407-TEE",
-            131_072,
-        ),
-        catalog_model(
-            "deepseek",
-            "deepseek-v4-flash",
-            "DeepSeek V4 Flash",
-            1_000_000,
-            384_000,
-            true,
-            false,
-        ),
-        catalog_model(
-            "deepseek",
-            "deepseek-v4-pro",
-            "DeepSeek V4 Pro",
-            1_000_000,
-            384_000,
-            true,
-            false,
-        ),
-        fireworks_model(
-            "accounts/fireworks/models/kimi-k2p6",
-            "Kimi K2.6",
-            Some(262_144),
-            true,
-            true,
-        ),
-        fireworks_model(
-            "accounts/fireworks/models/kimi-k2p7-code",
-            "Kimi K2.7 Code",
-            Some(262_144),
-            true,
-            true,
-        ),
-        fireworks_model(
-            "accounts/fireworks/models/deepseek-v4-flash",
-            "DeepSeek V4 Flash",
-            Some(1_048_576),
-            true,
-            false,
-        ),
-        fireworks_model(
-            "accounts/fireworks/models/deepseek-v4-pro",
-            "DeepSeek V4 Pro",
-            Some(1_048_576),
-            true,
-            false,
-        ),
-        fireworks_model(
-            "accounts/fireworks/models/glm-5p1",
-            "GLM 5.1",
-            Some(202_752),
-            true,
-            false,
-        ),
-        fireworks_model(
-            "accounts/fireworks/models/glm-5p2",
-            "GLM 5.2",
-            Some(1_048_576),
-            true,
-            false,
-        ),
-        fireworks_model(
-            "accounts/fireworks/models/gpt-oss-120b",
-            "GPT-OSS 120B",
-            Some(131_072),
-            true,
-            false,
-        ),
-        fireworks_model(
-            "accounts/fireworks/models/minimax-m2p7",
-            "MiniMax M2.7",
-            Some(196_608),
-            true,
-            false,
-        ),
-        fireworks_model(
-            "accounts/fireworks/models/minimax-m3",
-            "MiniMax M3",
-            Some(524_288),
-            true,
-            true,
-        ),
-        fireworks_model(
-            "accounts/fireworks/models/nemotron-3-ultra-nvfp4",
-            "Nemotron 3 Ultra NVFP4",
-            Some(262_144),
-            true,
-            false,
-        ),
-        fireworks_model(
-            "accounts/fireworks/models/qwen3p6-plus",
-            "Qwen 3.6 Plus",
-            None,
-            true,
-            true,
-        ),
-        fireworks_model(
-            "accounts/fireworks/models/qwen3p7-plus",
-            "Qwen 3.7 Plus",
-            Some(262_144),
-            true,
-            true,
-        ),
-        huggingface_gpt_oss_model(),
-        kilocode_auto_model(
-            "kilo-auto/frontier",
-            "Kilo Auto Frontier",
-            1_000_000,
-            128_000,
-            true,
-        ),
-        kilocode_auto_model(
-            "kilo-auto/balanced",
-            "Kilo Auto Balanced",
-            1_000_000,
-            65_536,
-            true,
-        ),
-        kilocode_auto_model(
-            "kilo-auto/efficient",
-            "Kilo Auto Efficient",
-            1_000_000,
-            65_536,
-            true,
-        ),
-        kilocode_auto_model(
-            "kilo-auto/free",
-            "Kilo Auto Free",
-            256_000,
-            10_000,
-            false,
-        ),
-        catalog_model(
-            "minimax",
-            "MiniMax-M3",
-            "MiniMax M3",
-            1_000_000,
-            32_768,
-            true,
-            true,
-        ),
-        catalog_model(
-            "minimax",
-            "MiniMax-M2.7",
-            "MiniMax M2.7",
-            204_800,
-            128_000,
-            true,
-            false,
-        ),
-        catalog_model(
-            "minimax",
-            "MiniMax-M2.7-highspeed",
-            "MiniMax M2.7 Highspeed",
-            204_800,
-            128_000,
-            true,
-            false,
-        ),
-        catalog_model(
-            "minimax",
-            "MiniMax-M2.5",
-            "MiniMax M2.5",
-            204_800,
-            128_000,
-            true,
-            false,
-        ),
-        catalog_model(
-            "minimax",
-            "MiniMax-M2.5-highspeed",
-            "MiniMax M2.5 Highspeed",
-            204_800,
-            128_000,
-            true,
-            false,
-        ),
-        catalog_model(
-            "minimax",
-            "MiniMax-M2.1",
-            "MiniMax M2.1",
-            204_800,
-            128_000,
-            true,
-            false,
-        ),
-        catalog_model(
-            "minimax",
-            "MiniMax-M2.1-highspeed",
-            "MiniMax M2.1 Highspeed",
-            204_800,
-            128_000,
-            true,
-            false,
-        ),
-        catalog_model(
-            "minimax",
-            "MiniMax-M2",
-            "MiniMax M2",
-            204_800,
-            128_000,
-            true,
-            false,
-        ),
-        catalog_model(
-            "mistral",
-            "codestral-latest",
-            "Codestral (latest)",
-            128_000,
-            4_096,
-            false,
-            false,
-        ),
-        catalog_model(
-            "mistral",
-            "mistral-large-latest",
-            "Mistral Large 3 (latest)",
-            256_000,
-            16_384,
-            false,
-            true,
-        ),
-        catalog_model(
-            "mistral",
-            "mistral-medium-latest",
-            "Mistral Medium 3.5 (latest)",
-            256_000,
-            8_192,
-            true,
-            true,
-        ),
-        catalog_model(
-            "mistral",
-            "mistral-small-latest",
-            "Mistral Small 4 (latest)",
-            256_000,
-            16_384,
-            false,
-            true,
-        ),
-        catalog_model(
-            "moonshot",
-            "kimi-k2.7-code",
-            "Kimi K2.7 Code",
-            262_144,
-            262_144,
-            true,
-            true,
-        ),
-        catalog_model(
-            "moonshot",
-            "kimi-k2.7-code-highspeed",
-            "Kimi K2.7 Code HighSpeed",
-            262_144,
-            262_144,
-            true,
-            true,
-        ),
-        catalog_model(
-            "moonshot",
-            "kimi-k2.6",
-            "Kimi K2.6",
-            262_144,
-            262_144,
-            true,
-            true,
-        ),
-        catalog_model(
-            "moonshot",
-            "kimi-k2.5",
-            "Kimi K2.5",
-            262_144,
-            262_144,
-            true,
-            true,
-        ),
-        catalog_model(
-            "moonshot",
-            "moonshot-v1-8k",
-            "Moonshot V1 8K",
-            8_192,
-            8_192,
-            false,
-            false,
-        ),
-        catalog_model(
-            "moonshot",
-            "moonshot-v1-32k",
-            "Moonshot V1 32K",
-            32_768,
-            32_768,
-            false,
-            false,
-        ),
-        catalog_model(
-            "moonshot",
-            "moonshot-v1-128k",
-            "Moonshot V1 128K",
-            131_072,
-            131_072,
-            false,
-            false,
-        ),
-        catalog_model(
-            "moonshot",
-            "moonshot-v1-auto",
-            "Moonshot V1 Auto",
-            131_072,
-            131_072,
-            false,
-            false,
-        ),
-        catalog_model(
-            "moonshot",
-            "moonshot-v1-8k-vision-preview",
-            "Moonshot V1 8K Vision Preview",
-            8_192,
-            8_192,
-            false,
-            true,
-        ),
-        catalog_model(
-            "moonshot",
-            "moonshot-v1-32k-vision-preview",
-            "Moonshot V1 32K Vision Preview",
-            32_768,
-            32_768,
-            false,
-            true,
-        ),
-        catalog_model(
-            "moonshot",
-            "moonshot-v1-128k-vision-preview",
-            "Moonshot V1 128K Vision Preview",
-            131_072,
-            131_072,
-            false,
-            true,
-        ),
-        catalog_model(
-            "nearai",
-            "zai-org/GLM-5.1-FP8",
-            "GLM 5.1 (NEAR AI Cloud TEE)",
-            202_752,
-            16_384,
-            true,
-            false,
-        ),
-        catalog_model(
-            "nearai",
-            "Qwen/Qwen3.6-35B-A3B-FP8",
-            "Qwen 3.6 35B A3B FP8 (NEAR AI Cloud TEE)",
-            262_144,
-            8_192,
-            true,
-            false,
-        ),
-        catalog_model(
-            "nearai",
-            "Qwen/Qwen3.5-122B-A10B",
-            "Qwen 3.5 122B A10B (NEAR AI Cloud TEE)",
-            262_144,
-            16_384,
-            true,
-            true,
-        ),
-        catalog_model(
-            "nearai",
-            "Qwen/Qwen3-VL-30B-A3B-Instruct",
-            "Qwen3 VL 30B A3B Instruct (NEAR AI Cloud TEE)",
-            16_384,
-            8_192,
-            false,
-            true,
-        ),
-        catalog_model(
-            "nearai",
-            "google/gemma-4-31B-it",
-            "Gemma 4 31B Instruct (NEAR AI Cloud TEE)",
-            262_144,
-            8_192,
-            true,
-            true,
-        ),
-        nvidia_model(
-            "nvidia/nemotron-3-super-120b-a12b",
-            "NVIDIA Nemotron 3 Super 120B",
-            1_000_000,
-            true,
-            false,
-        ),
-        nvidia_model(
-            "moonshotai/kimi-k2.6",
-            "Kimi K2.6",
-            262_144,
-            true,
-            true,
-        ),
-        nvidia_model(
-            "minimaxai/minimax-m2.7",
-            "MiniMax M2.7",
-            204_800,
-            true,
-            false,
-        ),
-        nvidia_model(
-            "minimaxai/minimax-m3",
-            "MiniMax M3",
-            1_000_000,
-            true,
-            true,
-        ),
-        nvidia_model("z-ai/glm-5.2", "GLM-5.2", 1_000_000, true, false),
-        opencode_go_model(
-            "deepseek-v4-pro",
-            "DeepSeek V4 Pro",
-            1_000_000,
-            Some(384_000),
-            true,
-            false,
-            None,
-        ),
-        opencode_go_model(
-            "deepseek-v4-flash",
-            "DeepSeek V4 Flash",
-            1_000_000,
-            Some(384_000),
-            true,
-            false,
-            None,
-        ),
-        opencode_go_model(
-            "glm-5.2",
-            "GLM-5.2",
-            1_000_000,
-            Some(131_072),
-            true,
-            false,
-            None,
-        ),
-        opencode_go_model(
-            "glm-5.1",
-            "GLM-5.1",
-            202_800,
-            Some(131_072),
-            true,
-            false,
-            None,
-        ),
-        opencode_go_model(
-            "kimi-k2.7-code",
-            "Kimi K2.7 Code",
-            262_144,
-            Some(262_144),
-            true,
-            true,
-            None,
-        ),
-        opencode_go_model(
-            "kimi-k2.6",
-            "Kimi K2.6",
-            262_144,
-            Some(262_144),
-            true,
-            true,
-            None,
-        ),
-        opencode_go_model(
-            "mimo-v2.5-pro",
-            "MiMo V2.5 Pro",
-            1_048_576,
-            Some(131_072),
-            true,
-            false,
-            None,
-        ),
-        opencode_go_model(
-            "mimo-v2.5",
-            "MiMo V2.5",
-            1_048_576,
-            Some(131_072),
-            true,
-            true,
-            None,
-        ),
-        opencode_go_model(
-            "minimax-m3",
-            "MiniMax M3",
-            1_000_000,
-            None,
-            true,
-            true,
-            Some("messages"),
-        ),
-        opencode_go_model(
-            "minimax-m2.7",
-            "MiniMax M2.7",
-            204_800,
-            None,
-            true,
-            false,
-            Some("messages"),
-        ),
-        opencode_go_model(
-            "minimax-m2.5",
-            "MiniMax M2.5",
-            196_608,
-            Some(32_768),
-            true,
-            false,
-            Some("messages"),
-        ),
-        opencode_go_model(
-            "qwen3.7-max",
-            "Qwen3.7 Max",
-            1_000_000,
-            Some(65_536),
-            true,
-            false,
-            Some("messages"),
-        ),
-        opencode_go_model(
-            "qwen3.7-plus",
-            "Qwen3.7 Plus",
-            1_000_000,
-            Some(65_536),
-            true,
-            true,
-            Some("messages"),
-        ),
-        opencode_go_model(
-            "qwen3.6-plus",
-            "Qwen3.6 Plus",
-            1_000_000,
-            Some(65_536),
-            true,
-            true,
-            Some("messages"),
-        ),
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(provider_id("openrouter"), "auto"),
-            display_name: "OpenRouter Auto".into(),
-            description: "OpenRouter Auto Router metadata aligned with the official Models API; the selected upstream model and provider are dynamic.".into(),
-            context_window_tokens: Some(2_000_000),
-            effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-            auto_compact_token_limit: None,
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: None,
-            tool_output_truncation_estimated_tokens: Some(
-                DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-            ),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::ConservativeBuiltin,
-            endpoint: None,
-        },
-        catalog_model(
-            "qianfan",
-            "deepseek-v3.2",
-            "DEEPSEEK V3.2",
-            131_072,
-            32_768,
-            false,
-            false,
-        ),
-        catalog_model(
-            "qianfan",
-            "deepseek-v3.2-think",
-            "DEEPSEEK V3.2 Think",
-            163_840,
-            65_536,
-            true,
-            false,
-        ),
-        catalog_model(
-            "qianfan",
-            "ernie-5.0",
-            "ERNIE 5.0",
-            248_832,
-            65_536,
-            false,
-            true,
-        ),
-        catalog_model(
-            "qianfan",
-            "ernie-5.0-thinking-preview",
-            "ERNIE 5.0 Thinking Preview",
-            248_832,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "qianfan",
-            "ernie-5.1",
-            "ERNIE 5.1",
-            248_832,
-            65_536,
-            false,
-            false,
-        ),
-        catalog_model(
-            "qianfan",
-            "ernie-x1.1",
-            "ERNIE X1.1",
-            121_856,
-            65_536,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "qwen3.7-plus",
-            "qwen3.7-plus",
-            1_000_000,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope",
-            "qwen3.7-plus-2026-05-26",
-            "qwen3.7-plus-2026-05-26",
-            1_000_000,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope",
-            "qwen3.7-max",
-            "qwen3.7-max",
-            1_000_000,
-            65_536,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "qwen3.7-max-2026-06-08",
-            "qwen3.7-max-2026-06-08",
-            1_000_000,
-            65_536,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "qwen3.7-max-2026-05-20",
-            "qwen3.7-max-2026-05-20",
-            1_000_000,
-            65_536,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "qwen3.5-plus",
-            "qwen3.5-plus",
-            1_000_000,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope",
-            "qwen3.6-plus",
-            "qwen3.6-plus",
-            1_000_000,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope",
-            "qwen3.6-flash",
-            "qwen3.6-flash",
-            1_000_000,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope",
-            "qwen3.5-flash",
-            "qwen3.5-flash",
-            1_000_000,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope",
-            "qwen3-max-2026-01-23",
-            "qwen3-max-2026-01-23",
-            262_144,
-            65_536,
-            false,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "qwen3-coder-next",
-            "qwen3-coder-next",
-            262_144,
-            65_536,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "qwen3-coder-plus",
-            "qwen3-coder-plus",
-            1_000_000,
-            65_536,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "qwen3-coder-flash",
-            "qwen3-coder-flash",
-            1_000_000,
-            65_536,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "qwen3.7-max",
-            "qwen3.7-max",
-            1_000_000,
-            65_536,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "qwen3.7-plus",
-            "qwen3.7-plus",
-            1_000_000,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "qwen3.6-plus",
-            "qwen3.6-plus",
-            1_000_000,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "qwen3.6-flash",
-            "qwen3.6-flash",
-            1_000_000,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "deepseek-v4-pro",
-            "DeepSeek V4 Pro",
-            1_000_000,
-            65_536, // Capped: DashScope gateway leaks </think> when max_output_tokens > 65536
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "deepseek-v4-flash",
-            "DeepSeek V4 Flash",
-            1_000_000,
-            65_536, // Capped: DashScope gateway leaks </think> when max_output_tokens > 65536
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "deepseek-v3.2",
-            "DeepSeek V3.2",
-            128_000,
-            32_768,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "kimi-k2.7-code",
-            "kimi-k2.7-code",
-            262_144,
-            65_536, // Capped: DashScope gateway leaks </think> when max_output_tokens > 65536
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "kimi-k2.6",
-            "kimi-k2.6",
-            262_144,
-            65_536, // Capped: DashScope gateway leaks </think> when max_output_tokens > 65536
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "kimi-k2.5",
-            "kimi-k2.5",
-            262_144,
-            32_768,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "glm-5.2",
-            "glm-5.2",
-            1_000_000,
-            65_536, // Capped: DashScope gateway leaks </think> when max_output_tokens > 65536
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "glm-5.1",
-            "glm-5.1",
-            202_752,
-            65_536, // Capped: DashScope gateway leaks </think> when max_output_tokens > 65536
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "glm-5",
-            "glm-5",
-            202_752,
-            16_384,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-token-plan",
-            "MiniMax-M2.5",
-            "MiniMax-M2.5",
-            196_608,
-            32_768,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-coding-plan",
-            "qwen3.7-plus",
-            "qwen3.7-plus",
-            1_000_000,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope-coding-plan",
-            "qwen3.6-plus",
-            "qwen3.6-plus",
-            1_000_000,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope-coding-plan",
-            "qwen3.5-plus",
-            "qwen3.5-plus",
-            1_000_000,
-            65_536,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope-coding-plan",
-            "qwen3-max-2026-01-23",
-            "qwen3-max-2026-01-23",
-            262_144,
-            65_536,
-            false,
-            false,
-        ),
-        catalog_model(
-            "dashscope-coding-plan",
-            "qwen3-coder-next",
-            "qwen3-coder-next",
-            262_144,
-            65_536,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-coding-plan",
-            "qwen3-coder-plus",
-            "qwen3-coder-plus",
-            1_000_000,
-            65_536,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-coding-plan",
-            "MiniMax-M2.5",
-            "MiniMax-M2.5",
-            196_608,
-            32_768,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-coding-plan",
-            "glm-5",
-            "glm-5",
-            202_752,
-            16_384,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-coding-plan",
-            "glm-4.7",
-            "glm-4.7",
-            202_752,
-            16_384,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope-coding-plan",
-            "kimi-k2.5",
-            "kimi-k2.5",
-            262_144,
-            32_768,
-            true,
-            true,
-        ),
-        stepfun_model("stepfun", "step-3.7-flash", "Step 3.7 Flash", true),
-        stepfun_model(
-            "stepfun",
-            "step-3.5-flash-2603",
-            "Step 3.5 Flash 2603",
-            false,
-        ),
-        stepfun_model("stepfun", "step-3.5-flash", "Step 3.5 Flash", false),
-        stepfun_model(
-            "stepfun-plan",
-            "step-3.7-flash",
-            "Step 3.7 Flash",
-            true,
-        ),
-        stepfun_model(
-            "stepfun-plan",
-            "step-3.5-flash-2603",
-            "Step 3.5 Flash 2603",
-            false,
-        ),
-        stepfun_model(
-            "stepfun-plan",
-            "step-3.5-flash",
-            "Step 3.5 Flash",
-            false,
-        ),
-        catalog_model("synthetic", "syn:large:text", "Synthetic Large Text", 524_288, 65_536, true, false),
-        catalog_model("synthetic", "syn:small:text", "Synthetic Small Text", 196_608, 65_536, true, false),
-        catalog_model("synthetic", "syn:large:vision", "Synthetic Large Vision", 262_144, 65_536, true, true),
-        catalog_model("synthetic", "syn:small:vision", "Synthetic Small Vision", 262_144, 65_536, true, true),
-        catalog_model("synthetic", "hf:openai/gpt-oss-120b", "OpenAI GPT OSS 120B", 131_072, 65_536, true, false),
-        catalog_model("synthetic", "hf:zai-org/GLM-5.2", "Z.ai GLM-5.2", 524_288, 65_536, true, false),
-        catalog_model("synthetic", "hf:moonshotai/Kimi-K2.7-Code", "Moonshot AI Kimi K2.7 Code", 262_144, 65_536, true, true),
-        catalog_model("synthetic", "hf:Qwen/Qwen3.6-27B", "Qwen3.6 27B", 262_144, 65_536, true, true),
-        catalog_model("synthetic", "hf:MiniMaxAI/MiniMax-M3", "MiniMax M3", 262_144, 65_536, true, true),
-        catalog_model("synthetic", "hf:zai-org/GLM-4.7-Flash", "Z.ai GLM-4.7 Flash", 196_608, 65_536, true, false),
-        catalog_model("synthetic", "hf:nvidia/NVIDIA-Nemotron-3-Super-120B-A12B-NVFP4", "NVIDIA Nemotron 3 Super 120B A12B NVFP4", 262_144, 65_536, true, false),
-        together_model(
-            "MiniMaxAI/MiniMax-M3",
-            "MiniMax M3",
-            524_288,
-            false,
-            true,
-        ),
-        together_model(
-            "MiniMaxAI/MiniMax-M2.7",
-            "MiniMax M2.7",
-            202_752,
-            true,
-            false,
-        ),
-        together_model(
-            "Qwen/Qwen3.5-9B",
-            "Qwen3.5 9B",
-            262_144,
-            true,
-            true,
-        ),
-        together_model(
-            "moonshotai/Kimi-K2.7-Code",
-            "Kimi K2.7 Code",
-            262_144,
-            false,
-            true,
-        ),
-        together_model(
-            "moonshotai/Kimi-K2.6",
-            "Kimi K2.6",
-            262_144,
-            true,
-            true,
-        ),
-        together_model("zai-org/GLM-5.2", "GLM-5.2", 262_144, false, false),
-        together_model(
-            "openai/gpt-oss-120b",
-            "GPT-OSS 120B",
-            128_000,
-            true,
-            false,
-        ),
-        together_model(
-            "deepseek-ai/DeepSeek-V4-Pro",
-            "DeepSeek V4 Pro",
-            512_000,
-            true,
-            false,
-        ),
-        together_model(
-            "nvidia/nemotron-3-ultra-550b-a55b",
-            "NVIDIA Nemotron 3 Ultra 550B",
-            512_300,
-            true,
-            false,
-        ),
-        together_model(
-            "meta-llama/Llama-3.3-70B-Instruct-Turbo",
-            "Llama 3.3 70B Instruct Turbo",
-            131_072,
-            false,
-            false,
-        ),
-        venice_model(
-            "zai-org-glm-4.7",
-            "GLM 4.7 (via Venice)",
-            198_000,
-            16_384,
-            true,
-            false,
-        ),
-        venice_model(
-            "qwen3-235b-a22b-thinking-2507",
-            "Qwen3 235B A22B Thinking 2507 (via Venice)",
-            128_000,
-            16_384,
-            true,
-            false,
-        ),
-        venice_model(
-            "qwen3-vl-235b-a22b",
-            "Qwen3 VL 235B (via Venice)",
-            128_000,
-            16_384,
-            false,
-            true,
-        ),
-        venice_model(
-            "qwen3-coder-480b-a35b-instruct-turbo",
-            "Qwen3 Coder 480B Turbo (via Venice)",
-            256_000,
-            65_536,
-            false,
-            false,
-        ),
-        venice_model(
-            "venice-uncensored-1-2",
-            "Venice Uncensored 1.2",
-            128_000,
-            8_192,
-            false,
-            true,
-        ),
-        catalog_model(
-            "vercel-ai-gateway",
-            "anthropic/claude-opus-4.6",
-            "Claude Opus 4.6",
-            1_000_000,
-            128_000,
-            true,
-            true,
-        ),
-        catalog_model(
-            "vercel-ai-gateway",
-            "openai/gpt-5.4",
-            "GPT 5.4",
-            1_050_000,
-            128_000,
-            true,
-            true,
-        ),
-        catalog_model(
-            "vercel-ai-gateway",
-            "openai/gpt-5.4-pro",
-            "GPT 5.4 Pro",
-            1_050_000,
-            128_000,
-            true,
-            true,
-        ),
-        catalog_model(
-            "vercel-ai-gateway",
-            "moonshotai/kimi-k2.6",
-            "Kimi K2.6",
-            262_000,
-            262_000,
-            true,
-            true,
-        ),
-        catalog_model(
-            "volcengine",
-            "doubao-seed-1-8-251228",
-            "Doubao Seed 1.8",
-            256_000,
-            4_096,
-            false,
-            true,
-        ),
-        catalog_model(
-            "volcengine",
-            "deepseek-v3-2-251201",
-            "DeepSeek V3.2",
-            128_000,
-            4_096,
-            false,
-            false,
-        ),
-        catalog_model(
-            "volcengine",
-            "doubao-seed-2-0-pro-260215",
-            "Doubao Seed 2.0 Pro",
-            256_000,
-            4_096,
-            true,
-            true,
-        ),
-        catalog_model(
-            "volcengine",
-            "doubao-seed-2-0-code-preview-260215",
-            "Doubao Seed 2.0 Code Preview",
-            256_000,
-            4_096,
-            false,
-            true,
-        ),
-        catalog_model(
-            "volcengine",
-            "doubao-seed-2-0-lite-260215",
-            "Doubao Seed 2.0 Lite",
-            256_000,
-            4_096,
-            false,
-            false,
-        ),
-        catalog_model(
-            "volcengine-coding",
-            "ark-code-latest",
-            "Ark Code Latest",
-            256_000,
-            65_536,
-            true,
-            false,
-        ),
-        catalog_model(
-            "volcengine-coding",
-            "doubao-seed-2-0-code-preview-260215",
-            "Doubao Seed 2.0 Code",
-            256_000,
-            4_096,
-            false,
-            false,
-        ),
-        catalog_model(
-            "volcengine-coding",
-            "doubao-seed-2-0-pro-260215",
-            "Doubao Seed 2.0 Pro",
-            256_000,
-            4_096,
-            false,
-            false,
-        ),
-        catalog_model(
-            "volcengine-coding",
-            "doubao-seed-2-0-lite-260215",
-            "Doubao Seed 2.0 Lite",
-            256_000,
-            4_096,
-            false,
-            false,
-        ),
-        catalog_model(
-            "volcengine-coding",
-            "deepseek-v3-2-251201",
-            "DeepSeek V3.2",
-            128_000,
-            4_096,
-            false,
-            false,
-        ),
-        catalog_model(
-            "volcengine-coding",
-            "deepseek-v4-pro",
-            "DeepSeek V4 Pro",
-            1_000_000,
-            8_192,
-            true,
-            false,
-        ),
-        catalog_model(
-            "volcengine-coding",
-            "deepseek-v4-flash",
-            "DeepSeek V4 Flash",
-            1_000_000,
-            8_192,
-            true,
-            false,
-        ),
-        catalog_model(
-            "volcengine-coding",
-            "kimi-k2.6",
-            "Kimi K2.6",
-            262_144,
-            32_768,
-            true,
-            false,
-        ),
-        catalog_model(
-            "volcengine-coding",
-            "kimi-k2.7-code",
-            "Kimi K2.7 Code",
-            262_144,
-            // Volcengine API rejects max_tokens > 32768 for kimi-k2.7-code
-            32_768,
-            true,
-            false,
-        ),
-        catalog_model(
-            "volcengine-coding",
-            "glm-5.2",
-            "GLM-5.2",
-            204_800,
-            128_000,
-            true,
-            false,
-        ),
-        catalog_model(
-            "volcengine-agent",
-            "ark-code-latest",
-            "Ark Code Latest",
-            256_000,
-            65_536,
-            true,
-            false,
-        ),
-        catalog_model(
-            "volcengine-agent",
-            "doubao-seed-2-0-code-preview-260215",
-            "Doubao Seed 2.0 Code",
-            256_000,
-            4_096,
-            false,
-            false,
-        ),
-        catalog_model(
-            "volcengine-agent",
-            "doubao-seed-2-0-pro-260215",
-            "Doubao Seed 2.0 Pro",
-            256_000,
-            4_096,
-            false,
-            false,
-        ),
-        catalog_model(
-            "volcengine-agent",
-            "doubao-seed-2-0-lite-260215",
-            "Doubao Seed 2.0 Lite",
-            256_000,
-            4_096,
-            false,
-            false,
-        ),
-        catalog_model(
-            "volcengine-agent",
-            "deepseek-v4-pro",
-            "DeepSeek V4 Pro",
-            1_000_000,
-            8_192,
-            true,
-            false,
-        ),
-        catalog_model(
-            "volcengine-agent",
-            "deepseek-v4-flash",
-            "DeepSeek V4 Flash",
-            1_000_000,
-            8_192,
-            true,
-            false,
-        ),
-        catalog_model(
-            "volcengine-agent",
-            "kimi-k2.6",
-            "Kimi K2.6",
-            262_144,
-            32_768,
-            true,
-            false,
-        ),
-        catalog_model(
-            "volcengine-agent",
-            "kimi-k2.7-code",
-            "Kimi K2.7 Code",
-            262_144,
-            // Volcengine API rejects max_tokens > 32768 for kimi-k2.7-code
-            32_768,
-            true,
-            false,
-        ),
-        catalog_model(
-            "volcengine-agent",
-            "glm-5.2",
-            "GLM-5.2",
-            204_800,
-            128_000,
-            true,
-            false,
-        ),
-        catalog_model(
-            "volcengine-agent",
-            "doubao-seed-2-0-mini-260215",
-            "Doubao Seed 2.0 Mini",
-            256_000,
-            4_096,
-            false,
-            false,
-        ),
-        catalog_model(
-            "volcengine-agent",
-            "MiniMax-M3",
-            "MiniMax M3",
-            1_000_000,
-            8_192,
-            true,
-            false,
-        ),
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(
-                provider_id("volcengine"),
-                "doubao-seedream-5.0-lite",
-            ),
-            display_name: "Doubao Seedream 5.0 Lite".into(),
-            description: "Volcengine Ark Seedream image generation model for the OpenAI Images-compatible API.".into(),
-            context_window_tokens: None,
-            effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
-            auto_compact_token_limit: None,
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: None,
-            tool_output_truncation_estimated_tokens: Some(
-                DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
-            ),
-            capabilities: ModelCapabilityFlags {
-                image_generation: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: Some(ProviderEndpointId::default_endpoint()),
-        },
-        catalog_model(
-            "xiaomi",
-            "mimo-v2.5-pro",
-            "Xiaomi MiMo V2.5 Pro",
-            1_048_576,
-            131_072,
-            true,
-            false,
-        ),
-        catalog_model(
-            "xiaomi",
-            "mimo-v2.5",
-            "Xiaomi MiMo V2.5",
-            1_048_576,
-            131_072,
-            true,
-            true,
-        ),
-        catalog_model(
-            "xiaomi-token-plan",
-            "mimo-v2.5-pro",
-            "Xiaomi MiMo V2.5 Pro",
-            1_048_576,
-            131_072,
-            true,
-            false,
-        ),
-        catalog_model(
-            "xiaomi-token-plan",
-            "mimo-v2.5",
-            "Xiaomi MiMo V2.5",
-            1_048_576,
-            131_072,
-            true,
-            true,
-        ),
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(provider_id("xai"), "grok-4.3"),
-            display_name: "Grok 4.3".into(),
-            description:
-                "xAI Grok 4.3 runtime defaults aligned with the official model page.".into(),
-            context_window_tokens: Some(1_000_000),
-            effective_context_window_percent: 90,
-            auto_compact_token_limit: Some(900_000),
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: None,
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        BuiltInModelMetadata {
-            model_ref: ModelRef::new(provider_id("xai"), "grok-4.5"),
-            display_name: "Grok 4.5".into(),
-            description:
-                "xAI Grok 4.5 runtime defaults aligned with the official model page.".into(),
-            context_window_tokens: Some(500_000),
-            effective_context_window_percent: 90,
-            auto_compact_token_limit: Some(450_000),
-            default_max_output_tokens: None,
-            max_output_tokens_upper_limit: None,
-            default_verbosity: None,
-            tool_output_truncation_estimated_tokens: Some(2_500),
-            capabilities: ModelCapabilityFlags {
-                image_input: true,
-                supports_reasoning: true,
-                ..ModelCapabilityFlags::default()
-            },
-            reasoning_effort_options: Vec::new(),
-            source: ModelMetadataSource::BuiltInCatalog,
-            endpoint: None,
-        },
-        catalog_model("zai", "glm-5.2", "GLM-5.2", 1_000_000, 131_072, true, false),
-        catalog_model("zai", "glm-5.1", "GLM-5.1", 202_800, 131_072, true, false),
-        catalog_model("zai", "glm-5", "GLM-5", 202_800, 131_072, true, false),
-        catalog_model(
-            "zai",
-            "glm-5-turbo",
-            "GLM-5 Turbo",
-            202_800,
-            131_072,
-            true,
-            false,
-        ),
-        catalog_model(
-            "zai",
-            "glm-5v-turbo",
-            "GLM-5V Turbo",
-            202_800,
-            131_072,
-            true,
-            true,
-        ),
-        catalog_model("zai", "glm-4.7", "GLM-4.7", 204_800, 131_072, true, false),
-        catalog_model(
-            "zai",
-            "glm-4.7-flash",
-            "GLM-4.7 Flash",
-            200_000,
-            131_072,
-            true,
-            false,
-        ),
-        catalog_model(
-            "zai",
-            "glm-4.7-flashx",
-            "GLM-4.7 FlashX",
-            200_000,
-            128_000,
-            true,
-            false,
-        ),
-        catalog_model("zai", "glm-4.6", "GLM-4.6", 204_800, 131_072, true, false),
-        catalog_model("zai", "glm-4.6v", "GLM-4.6V", 128_000, 32_768, true, true),
-        catalog_model("zai", "glm-4.5", "GLM-4.5", 131_072, 98_304, true, false),
-        catalog_model(
-            "zai",
-            "glm-4.5-air",
-            "GLM-4.5 Air",
-            131_072,
-            98_304,
-            true,
-            false,
-        ),
-        catalog_model(
-            "zai",
-            "glm-4.5-x",
-            "GLM-4.5 X",
-            131_072,
-            98_304,
-            true,
-            false,
-        ),
-        catalog_model(
-            "zai",
-            "glm-4.5-airx",
-            "GLM-4.5 AirX",
-            131_072,
-            98_304,
-            true,
-            false,
-        ),
-        catalog_model(
-            "zai",
-            "glm-4.5-flash",
-            "GLM-4.5 Flash",
-            131_072,
-            98_304,
-            true,
-            false,
-        ),
-        catalog_model("zai", "glm-4.5v", "GLM-4.5V", 64_000, 16_384, true, true),
-        catalog_model(
-            "zai",
-            "glm-4.6v-flashx",
-            "GLM-4.6V FlashX",
-            128_000,
-            32_768,
-            true,
-            true,
-        ),
-        catalog_model(
-            "zai",
-            "glm-4.6v-flash",
-            "GLM-4.6V Flash",
-            128_000,
-            32_768,
-            true,
-            true,
-        ),
-        catalog_model(
-            "zai",
-            "glm-4-32b-0414-128k",
-            "GLM-4 32B 0414 128K",
-            131_072,
-            16_384,
-            false,
-            false,
-        ),
-    ];
-    entries.extend([
-        catalog_model(
-            "bigmodel", "glm-5.2", "GLM-5.2", 1_000_000, 131_072, true, false,
-        ),
-        catalog_model(
-            "bigmodel", "glm-5.1", "GLM-5.1", 204_800, 131_072, true, false,
-        ),
-        catalog_model("bigmodel", "glm-5", "GLM-5", 204_800, 131_072, true, false),
-        catalog_model(
-            "bigmodel",
-            "glm-5-turbo",
-            "GLM-5 Turbo",
-            204_800,
-            131_072,
-            true,
-            false,
-        ),
-        catalog_model(
-            "bigmodel",
-            "glm-5v-turbo",
-            "GLM-5V Turbo",
-            204_800,
-            131_072,
-            true,
-            true,
-        ),
-        catalog_model(
-            "bigmodel", "glm-4.7", "GLM-4.7", 204_800, 131_072, true, false,
-        ),
-        catalog_model(
-            "bigmodel",
-            "glm-4.7-flashx",
-            "GLM-4.7 FlashX",
-            204_800,
-            131_072,
-            true,
-            false,
-        ),
-        catalog_model(
-            "bigmodel",
-            "glm-4.7-flash",
-            "GLM-4.7 Flash",
-            204_800,
-            131_072,
-            true,
-            false,
-        ),
-        catalog_model(
-            "bigmodel", "glm-4.6", "GLM-4.6", 204_800, 131_072, true, false,
-        ),
-        catalog_model(
-            "bigmodel",
-            "glm-4.5-air",
-            "GLM-4.5 Air",
-            131_072,
-            98_304,
-            true,
-            false,
-        ),
-        catalog_model(
-            "bigmodel",
-            "glm-4.5-airx",
-            "GLM-4.5 AirX",
-            131_072,
-            98_304,
-            true,
-            false,
-        ),
-        catalog_model(
-            "bigmodel",
-            "glm-4.5-flash",
-            "GLM-4.5 Flash",
-            131_072,
-            98_304,
-            true,
-            false,
-        ),
-        catalog_model(
-            "bigmodel",
-            "glm-4-long",
-            "GLM-4 Long",
-            1_000_000,
-            4_096,
-            false,
-            false,
-        ),
-        catalog_model(
-            "bigmodel",
-            "glm-4-flashx-250414",
-            "GLM-4 FlashX 250414",
-            131_072,
-            16_384,
-            false,
-            false,
-        ),
-        catalog_model(
-            "bigmodel",
-            "glm-4-flash-250414",
-            "GLM-4 Flash 250414",
-            131_072,
-            16_384,
-            false,
-            false,
-        ),
-        catalog_model(
-            "bigmodel", "glm-4.6v", "GLM-4.6V", 131_072, 32_768, true, true,
-        ),
-        catalog_model(
-            "bigmodel",
-            "glm-4.6v-flash",
-            "GLM-4.6V Flash",
-            131_072,
-            32_768,
-            true,
-            true,
-        ),
-        catalog_model(
-            "bigmodel",
-            "glm-4.1v-thinking-flashx",
-            "GLM-4.1V Thinking FlashX",
-            65_536,
-            16_384,
-            true,
-            true,
-        ),
-        catalog_model(
-            "bigmodel",
-            "glm-4.1v-thinking-flash",
-            "GLM-4.1V Thinking Flash",
-            65_536,
-            16_384,
-            true,
-            true,
-        ),
-        catalog_model(
-            "bigmodel",
-            "glm-4v-flash",
-            "GLM-4V Flash",
-            16_384,
-            1_024,
-            false,
-            true,
-        ),
-        catalog_model(
-            "dashscope",
-            "deepseek-v4-pro",
-            "DeepSeek V4 Pro",
-            1_000_000,
-            384_000,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "deepseek-v4-flash",
-            "DeepSeek V4 Flash",
-            1_000_000,
-            384_000,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "ZHIPU/GLM-5.2",
-            "ZHIPU/GLM-5.2",
-            1_000_000,
-            131_072,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "glm-5.2",
-            "glm-5.2",
-            204_800,
-            128_000,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "glm-5.1",
-            "glm-5.1",
-            202_752,
-            131_072,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "kimi-k2.7-code",
-            "kimi-k2.7-code",
-            262_144,
-            98_304,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope",
-            "kimi-k2.6",
-            "kimi-k2.6",
-            262_144,
-            98_304,
-            true,
-            true,
-        ),
-        catalog_model(
-            "dashscope",
-            "MiniMax/MiniMax-M3",
-            "MiniMax/MiniMax-M3",
-            196_608,
-            32_768,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "MiniMax-M2.5",
-            "MiniMax-M2.5",
-            196_608,
-            32_768,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "mimo-v2.5-pro",
-            "MiMo V2.5 Pro",
-            1_000_000,
-            131_072,
-            true,
-            false,
-        ),
-        catalog_model("dashscope", "glm-5", "glm-5", 202_752, 16_384, true, false),
-        catalog_model(
-            "dashscope",
-            "glm-4.7",
-            "glm-4.7",
-            202_752,
-            16_384,
-            true,
-            false,
-        ),
-        catalog_model(
-            "dashscope",
-            "kimi-k2.5",
-            "kimi-k2.5",
-            262_144,
-            32_768,
-            true,
-            true,
-        ),
-    ]);
-    entries.extend(tencent_tokenhub_model_entries());
-    entries
 }
 
 #[cfg(test)]
