+++ src/model_catalog/providers/china.rs
@@ -1,0 +1,1145 @@
+use super::super::*;
+use super::common::catalog_model;
+
+pub(super) fn stepfun_model(
+    provider: &str,
+    model: &str,
+    display_name: &str,
+    image_input: bool,
+) -> BuiltInModelMetadata {
+    let model_ref = ModelRef::new(provider_id(provider), model);
+    BuiltInModelMetadata {
+        default_verbosity: default_verbosity_for_model(&model_ref),
+        model_ref,
+        display_name: display_name.into(),
+        description: format!("Holon built-in runtime metadata for the StepFun {model} model."),
+        context_window_tokens: Some(262_144),
+        effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
+        auto_compact_token_limit: None,
+        default_max_output_tokens: None,
+        max_output_tokens_upper_limit: None,
+        tool_output_truncation_estimated_tokens: Some(
+            DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
+        ),
+        capabilities: ModelCapabilityFlags {
+            image_input,
+            supports_reasoning: true,
+            ..ModelCapabilityFlags::default()
+        },
+        reasoning_effort_options: Vec::new(),
+        source: ModelMetadataSource::BuiltInCatalog,
+        endpoint: None,
+    }
+}
+
+pub(super) fn early_entries() -> Vec<BuiltInModelMetadata> {
+    vec![
+        catalog_model(
+            "qianfan",
+            "deepseek-v3.2",
+            "DEEPSEEK V3.2",
+            131_072,
+            32_768,
+            false,
+            false,
+        ),
+        catalog_model(
+            "qianfan",
+            "deepseek-v3.2-think",
+            "DEEPSEEK V3.2 Think",
+            163_840,
+            65_536,
+            true,
+            false,
+        ),
+        catalog_model(
+            "qianfan",
+            "ernie-5.0",
+            "ERNIE 5.0",
+            248_832,
+            65_536,
+            false,
+            true,
+        ),
+        catalog_model(
+            "qianfan",
+            "ernie-5.0-thinking-preview",
+            "ERNIE 5.0 Thinking Preview",
+            248_832,
+            65_536,
+            true,
+            true,
+        ),
+        catalog_model(
+            "qianfan",
+            "ernie-5.1",
+            "ERNIE 5.1",
+            248_832,
+            65_536,
+            false,
+            false,
+        ),
+        catalog_model(
+            "qianfan",
+            "ernie-x1.1",
+            "ERNIE X1.1",
+            121_856,
+            65_536,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "qwen3.7-plus",
+            "qwen3.7-plus",
+            1_000_000,
+            65_536,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope",
+            "qwen3.7-plus-2026-05-26",
+            "qwen3.7-plus-2026-05-26",
+            1_000_000,
+            65_536,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope",
+            "qwen3.7-max",
+            "qwen3.7-max",
+            1_000_000,
+            65_536,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "qwen3.7-max-2026-06-08",
+            "qwen3.7-max-2026-06-08",
+            1_000_000,
+            65_536,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "qwen3.7-max-2026-05-20",
+            "qwen3.7-max-2026-05-20",
+            1_000_000,
+            65_536,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "qwen3.5-plus",
+            "qwen3.5-plus",
+            1_000_000,
+            65_536,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope",
+            "qwen3.6-plus",
+            "qwen3.6-plus",
+            1_000_000,
+            65_536,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope",
+            "qwen3.6-flash",
+            "qwen3.6-flash",
+            1_000_000,
+            65_536,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope",
+            "qwen3.5-flash",
+            "qwen3.5-flash",
+            1_000_000,
+            65_536,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope",
+            "qwen3-max-2026-01-23",
+            "qwen3-max-2026-01-23",
+            262_144,
+            65_536,
+            false,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "qwen3-coder-next",
+            "qwen3-coder-next",
+            262_144,
+            65_536,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "qwen3-coder-plus",
+            "qwen3-coder-plus",
+            1_000_000,
+            65_536,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "qwen3-coder-flash",
+            "qwen3-coder-flash",
+            1_000_000,
+            65_536,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "qwen3.7-max",
+            "qwen3.7-max",
+            1_000_000,
+            65_536,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "qwen3.7-plus",
+            "qwen3.7-plus",
+            1_000_000,
+            65_536,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "qwen3.6-plus",
+            "qwen3.6-plus",
+            1_000_000,
+            65_536,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "qwen3.6-flash",
+            "qwen3.6-flash",
+            1_000_000,
+            65_536,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "deepseek-v4-pro",
+            "DeepSeek V4 Pro",
+            1_000_000,
+            65_536, // Capped: DashScope gateway leaks </think> when max_output_tokens > 65536
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "deepseek-v4-flash",
+            "DeepSeek V4 Flash",
+            1_000_000,
+            65_536, // Capped: DashScope gateway leaks </think> when max_output_tokens > 65536
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "deepseek-v3.2",
+            "DeepSeek V3.2",
+            128_000,
+            32_768,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "kimi-k2.7-code",
+            "kimi-k2.7-code",
+            262_144,
+            65_536, // Capped: DashScope gateway leaks </think> when max_output_tokens > 65536
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "kimi-k2.6",
+            "kimi-k2.6",
+            262_144,
+            65_536, // Capped: DashScope gateway leaks </think> when max_output_tokens > 65536
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "kimi-k2.5",
+            "kimi-k2.5",
+            262_144,
+            32_768,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "glm-5.2",
+            "glm-5.2",
+            1_000_000,
+            65_536, // Capped: DashScope gateway leaks </think> when max_output_tokens > 65536
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "glm-5.1",
+            "glm-5.1",
+            202_752,
+            65_536, // Capped: DashScope gateway leaks </think> when max_output_tokens > 65536
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "glm-5",
+            "glm-5",
+            202_752,
+            16_384,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-token-plan",
+            "MiniMax-M2.5",
+            "MiniMax-M2.5",
+            196_608,
+            32_768,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-coding-plan",
+            "qwen3.7-plus",
+            "qwen3.7-plus",
+            1_000_000,
+            65_536,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope-coding-plan",
+            "qwen3.6-plus",
+            "qwen3.6-plus",
+            1_000_000,
+            65_536,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope-coding-plan",
+            "qwen3.5-plus",
+            "qwen3.5-plus",
+            1_000_000,
+            65_536,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope-coding-plan",
+            "qwen3-max-2026-01-23",
+            "qwen3-max-2026-01-23",
+            262_144,
+            65_536,
+            false,
+            false,
+        ),
+        catalog_model(
+            "dashscope-coding-plan",
+            "qwen3-coder-next",
+            "qwen3-coder-next",
+            262_144,
+            65_536,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-coding-plan",
+            "qwen3-coder-plus",
+            "qwen3-coder-plus",
+            1_000_000,
+            65_536,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-coding-plan",
+            "MiniMax-M2.5",
+            "MiniMax-M2.5",
+            196_608,
+            32_768,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-coding-plan",
+            "glm-5",
+            "glm-5",
+            202_752,
+            16_384,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-coding-plan",
+            "glm-4.7",
+            "glm-4.7",
+            202_752,
+            16_384,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope-coding-plan",
+            "kimi-k2.5",
+            "kimi-k2.5",
+            262_144,
+            32_768,
+            true,
+            true,
+        ),
+        stepfun_model("stepfun", "step-3.7-flash", "Step 3.7 Flash", true),
+        stepfun_model(
+            "stepfun",
+            "step-3.5-flash-2603",
+            "Step 3.5 Flash 2603",
+            false,
+        ),
+        stepfun_model("stepfun", "step-3.5-flash", "Step 3.5 Flash", false),
+        stepfun_model("stepfun-plan", "step-3.7-flash", "Step 3.7 Flash", true),
+        stepfun_model(
+            "stepfun-plan",
+            "step-3.5-flash-2603",
+            "Step 3.5 Flash 2603",
+            false,
+        ),
+        stepfun_model("stepfun-plan", "step-3.5-flash", "Step 3.5 Flash", false),
+    ]
+}
+
+pub(super) fn late_entries() -> Vec<BuiltInModelMetadata> {
+    let mut entries = vec![
+        catalog_model(
+            "volcengine",
+            "doubao-seed-1-8-251228",
+            "Doubao Seed 1.8",
+            256_000,
+            4_096,
+            false,
+            true,
+        ),
+        catalog_model(
+            "volcengine",
+            "deepseek-v3-2-251201",
+            "DeepSeek V3.2",
+            128_000,
+            4_096,
+            false,
+            false,
+        ),
+        catalog_model(
+            "volcengine",
+            "doubao-seed-2-0-pro-260215",
+            "Doubao Seed 2.0 Pro",
+            256_000,
+            4_096,
+            true,
+            true,
+        ),
+        catalog_model(
+            "volcengine",
+            "doubao-seed-2-0-code-preview-260215",
+            "Doubao Seed 2.0 Code Preview",
+            256_000,
+            4_096,
+            false,
+            true,
+        ),
+        catalog_model(
+            "volcengine",
+            "doubao-seed-2-0-lite-260215",
+            "Doubao Seed 2.0 Lite",
+            256_000,
+            4_096,
+            false,
+            false,
+        ),
+        catalog_model(
+            "volcengine-coding",
+            "ark-code-latest",
+            "Ark Code Latest",
+            256_000,
+            65_536,
+            true,
+            false,
+        ),
+        catalog_model(
+            "volcengine-coding",
+            "doubao-seed-2-0-code-preview-260215",
+            "Doubao Seed 2.0 Code",
+            256_000,
+            4_096,
+            false,
+            false,
+        ),
+        catalog_model(
+            "volcengine-coding",
+            "doubao-seed-2-0-pro-260215",
+            "Doubao Seed 2.0 Pro",
+            256_000,
+            4_096,
+            false,
+            false,
+        ),
+        catalog_model(
+            "volcengine-coding",
+            "doubao-seed-2-0-lite-260215",
+            "Doubao Seed 2.0 Lite",
+            256_000,
+            4_096,
+            false,
+            false,
+        ),
+        catalog_model(
+            "volcengine-coding",
+            "deepseek-v3-2-251201",
+            "DeepSeek V3.2",
+            128_000,
+            4_096,
+            false,
+            false,
+        ),
+        catalog_model(
+            "volcengine-coding",
+            "deepseek-v4-pro",
+            "DeepSeek V4 Pro",
+            1_000_000,
+            8_192,
+            true,
+            false,
+        ),
+        catalog_model(
+            "volcengine-coding",
+            "deepseek-v4-flash",
+            "DeepSeek V4 Flash",
+            1_000_000,
+            8_192,
+            true,
+            false,
+        ),
+        catalog_model(
+            "volcengine-coding",
+            "kimi-k2.6",
+            "Kimi K2.6",
+            262_144,
+            32_768,
+            true,
+            false,
+        ),
+        catalog_model(
+            "volcengine-coding",
+            "kimi-k2.7-code",
+            "Kimi K2.7 Code",
+            262_144,
+            // Volcengine API rejects max_tokens > 32768 for kimi-k2.7-code
+            32_768,
+            true,
+            false,
+        ),
+        catalog_model(
+            "volcengine-coding",
+            "glm-5.2",
+            "GLM-5.2",
+            204_800,
+            128_000,
+            true,
+            false,
+        ),
+        catalog_model(
+            "volcengine-agent",
+            "ark-code-latest",
+            "Ark Code Latest",
+            256_000,
+            65_536,
+            true,
+            false,
+        ),
+        catalog_model(
+            "volcengine-agent",
+            "doubao-seed-2-0-code-preview-260215",
+            "Doubao Seed 2.0 Code",
+            256_000,
+            4_096,
+            false,
+            false,
+        ),
+        catalog_model(
+            "volcengine-agent",
+            "doubao-seed-2-0-pro-260215",
+            "Doubao Seed 2.0 Pro",
+            256_000,
+            4_096,
+            false,
+            false,
+        ),
+        catalog_model(
+            "volcengine-agent",
+            "doubao-seed-2-0-lite-260215",
+            "Doubao Seed 2.0 Lite",
+            256_000,
+            4_096,
+            false,
+            false,
+        ),
+        catalog_model(
+            "volcengine-agent",
+            "deepseek-v4-pro",
+            "DeepSeek V4 Pro",
+            1_000_000,
+            8_192,
+            true,
+            false,
+        ),
+        catalog_model(
+            "volcengine-agent",
+            "deepseek-v4-flash",
+            "DeepSeek V4 Flash",
+            1_000_000,
+            8_192,
+            true,
+            false,
+        ),
+        catalog_model(
+            "volcengine-agent",
+            "kimi-k2.6",
+            "Kimi K2.6",
+            262_144,
+            32_768,
+            true,
+            false,
+        ),
+        catalog_model(
+            "volcengine-agent",
+            "kimi-k2.7-code",
+            "Kimi K2.7 Code",
+            262_144,
+            // Volcengine API rejects max_tokens > 32768 for kimi-k2.7-code
+            32_768,
+            true,
+            false,
+        ),
+        catalog_model(
+            "volcengine-agent",
+            "glm-5.2",
+            "GLM-5.2",
+            204_800,
+            128_000,
+            true,
+            false,
+        ),
+        catalog_model(
+            "volcengine-agent",
+            "doubao-seed-2-0-mini-260215",
+            "Doubao Seed 2.0 Mini",
+            256_000,
+            4_096,
+            false,
+            false,
+        ),
+        catalog_model(
+            "volcengine-agent",
+            "MiniMax-M3",
+            "MiniMax M3",
+            1_000_000,
+            8_192,
+            true,
+            false,
+        ),
+        BuiltInModelMetadata {
+            model_ref: ModelRef::new(
+                provider_id("volcengine"),
+                "doubao-seedream-5.0-lite",
+            ),
+            display_name: "Doubao Seedream 5.0 Lite".into(),
+            description: "Volcengine Ark Seedream image generation model for the OpenAI Images-compatible API.".into(),
+            context_window_tokens: None,
+            effective_context_window_percent: DEFAULT_EFFECTIVE_CONTEXT_WINDOW_PERCENT,
+            auto_compact_token_limit: None,
+            default_max_output_tokens: None,
+            max_output_tokens_upper_limit: None,
+            default_verbosity: None,
+            tool_output_truncation_estimated_tokens: Some(
+                DEFAULT_TOOL_OUTPUT_TRUNCATION_ESTIMATED_TOKENS,
+            ),
+            capabilities: ModelCapabilityFlags {
+                image_generation: true,
+                ..ModelCapabilityFlags::default()
+            },
+            reasoning_effort_options: Vec::new(),
+            source: ModelMetadataSource::BuiltInCatalog,
+            endpoint: Some(ProviderEndpointId::default_endpoint()),
+        },
+        catalog_model(
+            "xiaomi",
+            "mimo-v2.5-pro",
+            "Xiaomi MiMo V2.5 Pro",
+            1_048_576,
+            131_072,
+            true,
+            false,
+        ),
+        catalog_model(
+            "xiaomi",
+            "mimo-v2.5",
+            "Xiaomi MiMo V2.5",
+            1_048_576,
+            131_072,
+            true,
+            true,
+        ),
+        catalog_model(
+            "xiaomi-token-plan",
+            "mimo-v2.5-pro",
+            "Xiaomi MiMo V2.5 Pro",
+            1_048_576,
+            131_072,
+            true,
+            false,
+        ),
+        catalog_model(
+            "xiaomi-token-plan",
+            "mimo-v2.5",
+            "Xiaomi MiMo V2.5",
+            1_048_576,
+            131_072,
+            true,
+            true,
+        ),
+        BuiltInModelMetadata {
+            model_ref: ModelRef::new(provider_id("xai"), "grok-4.3"),
+            display_name: "Grok 4.3".into(),
+            description:
+                "xAI Grok 4.3 runtime defaults aligned with the official model page.".into(),
+            context_window_tokens: Some(1_000_000),
+            effective_context_window_percent: 90,
+            auto_compact_token_limit: Some(900_000),
+            default_max_output_tokens: None,
+            max_output_tokens_upper_limit: None,
+            default_verbosity: None,
+            tool_output_truncation_estimated_tokens: Some(2_500),
+            capabilities: ModelCapabilityFlags {
+                image_input: true,
+                supports_reasoning: true,
+                ..ModelCapabilityFlags::default()
+            },
+            reasoning_effort_options: Vec::new(),
+            source: ModelMetadataSource::BuiltInCatalog,
+            endpoint: None,
+        },
+        BuiltInModelMetadata {
+            model_ref: ModelRef::new(provider_id("xai"), "grok-4.5"),
+            display_name: "Grok 4.5".into(),
+            description:
+                "xAI Grok 4.5 runtime defaults aligned with the official model page.".into(),
+            context_window_tokens: Some(500_000),
+            effective_context_window_percent: 90,
+            auto_compact_token_limit: Some(450_000),
+            default_max_output_tokens: None,
+            max_output_tokens_upper_limit: None,
+            default_verbosity: None,
+            tool_output_truncation_estimated_tokens: Some(2_500),
+            capabilities: ModelCapabilityFlags {
+                image_input: true,
+                supports_reasoning: true,
+                ..ModelCapabilityFlags::default()
+            },
+            reasoning_effort_options: Vec::new(),
+            source: ModelMetadataSource::BuiltInCatalog,
+            endpoint: None,
+        },
+        catalog_model("zai", "glm-5.2", "GLM-5.2", 1_000_000, 131_072, true, false),
+        catalog_model("zai", "glm-5.1", "GLM-5.1", 202_800, 131_072, true, false),
+        catalog_model("zai", "glm-5", "GLM-5", 202_800, 131_072, true, false),
+        catalog_model(
+            "zai",
+            "glm-5-turbo",
+            "GLM-5 Turbo",
+            202_800,
+            131_072,
+            true,
+            false,
+        ),
+        catalog_model(
+            "zai",
+            "glm-5v-turbo",
+            "GLM-5V Turbo",
+            202_800,
+            131_072,
+            true,
+            true,
+        ),
+        catalog_model("zai", "glm-4.7", "GLM-4.7", 204_800, 131_072, true, false),
+        catalog_model(
+            "zai",
+            "glm-4.7-flash",
+            "GLM-4.7 Flash",
+            200_000,
+            131_072,
+            true,
+            false,
+        ),
+        catalog_model(
+            "zai",
+            "glm-4.7-flashx",
+            "GLM-4.7 FlashX",
+            200_000,
+            128_000,
+            true,
+            false,
+        ),
+        catalog_model("zai", "glm-4.6", "GLM-4.6", 204_800, 131_072, true, false),
+        catalog_model("zai", "glm-4.6v", "GLM-4.6V", 128_000, 32_768, true, true),
+        catalog_model("zai", "glm-4.5", "GLM-4.5", 131_072, 98_304, true, false),
+        catalog_model(
+            "zai",
+            "glm-4.5-air",
+            "GLM-4.5 Air",
+            131_072,
+            98_304,
+            true,
+            false,
+        ),
+        catalog_model(
+            "zai",
+            "glm-4.5-x",
+            "GLM-4.5 X",
+            131_072,
+            98_304,
+            true,
+            false,
+        ),
+        catalog_model(
+            "zai",
+            "glm-4.5-airx",
+            "GLM-4.5 AirX",
+            131_072,
+            98_304,
+            true,
+            false,
+        ),
+        catalog_model(
+            "zai",
+            "glm-4.5-flash",
+            "GLM-4.5 Flash",
+            131_072,
+            98_304,
+            true,
+            false,
+        ),
+        catalog_model("zai", "glm-4.5v", "GLM-4.5V", 64_000, 16_384, true, true),
+        catalog_model(
+            "zai",
+            "glm-4.6v-flashx",
+            "GLM-4.6V FlashX",
+            128_000,
+            32_768,
+            true,
+            true,
+        ),
+        catalog_model(
+            "zai",
+            "glm-4.6v-flash",
+            "GLM-4.6V Flash",
+            128_000,
+            32_768,
+            true,
+            true,
+        ),
+        catalog_model(
+            "zai",
+            "glm-4-32b-0414-128k",
+            "GLM-4 32B 0414 128K",
+            131_072,
+            16_384,
+            false,
+            false,
+        ),
+    ];
+    entries.extend([
+        catalog_model(
+            "bigmodel", "glm-5.2", "GLM-5.2", 1_000_000, 131_072, true, false,
+        ),
+        catalog_model(
+            "bigmodel", "glm-5.1", "GLM-5.1", 204_800, 131_072, true, false,
+        ),
+        catalog_model("bigmodel", "glm-5", "GLM-5", 204_800, 131_072, true, false),
+        catalog_model(
+            "bigmodel",
+            "glm-5-turbo",
+            "GLM-5 Turbo",
+            204_800,
+            131_072,
+            true,
+            false,
+        ),
+        catalog_model(
+            "bigmodel",
+            "glm-5v-turbo",
+            "GLM-5V Turbo",
+            204_800,
+            131_072,
+            true,
+            true,
+        ),
+        catalog_model(
+            "bigmodel", "glm-4.7", "GLM-4.7", 204_800, 131_072, true, false,
+        ),
+        catalog_model(
+            "bigmodel",
+            "glm-4.7-flashx",
+            "GLM-4.7 FlashX",
+            204_800,
+            131_072,
+            true,
+            false,
+        ),
+        catalog_model(
+            "bigmodel",
+            "glm-4.7-flash",
+            "GLM-4.7 Flash",
+            204_800,
+            131_072,
+            true,
+            false,
+        ),
+        catalog_model(
+            "bigmodel", "glm-4.6", "GLM-4.6", 204_800, 131_072, true, false,
+        ),
+        catalog_model(
+            "bigmodel",
+            "glm-4.5-air",
+            "GLM-4.5 Air",
+            131_072,
+            98_304,
+            true,
+            false,
+        ),
+        catalog_model(
+            "bigmodel",
+            "glm-4.5-airx",
+            "GLM-4.5 AirX",
+            131_072,
+            98_304,
+            true,
+            false,
+        ),
+        catalog_model(
+            "bigmodel",
+            "glm-4.5-flash",
+            "GLM-4.5 Flash",
+            131_072,
+            98_304,
+            true,
+            false,
+        ),
+        catalog_model(
+            "bigmodel",
+            "glm-4-long",
+            "GLM-4 Long",
+            1_000_000,
+            4_096,
+            false,
+            false,
+        ),
+        catalog_model(
+            "bigmodel",
+            "glm-4-flashx-250414",
+            "GLM-4 FlashX 250414",
+            131_072,
+            16_384,
+            false,
+            false,
+        ),
+        catalog_model(
+            "bigmodel",
+            "glm-4-flash-250414",
+            "GLM-4 Flash 250414",
+            131_072,
+            16_384,
+            false,
+            false,
+        ),
+        catalog_model(
+            "bigmodel", "glm-4.6v", "GLM-4.6V", 131_072, 32_768, true, true,
+        ),
+        catalog_model(
+            "bigmodel",
+            "glm-4.6v-flash",
+            "GLM-4.6V Flash",
+            131_072,
+            32_768,
+            true,
+            true,
+        ),
+        catalog_model(
+            "bigmodel",
+            "glm-4.1v-thinking-flashx",
+            "GLM-4.1V Thinking FlashX",
+            65_536,
+            16_384,
+            true,
+            true,
+        ),
+        catalog_model(
+            "bigmodel",
+            "glm-4.1v-thinking-flash",
+            "GLM-4.1V Thinking Flash",
+            65_536,
+            16_384,
+            true,
+            true,
+        ),
+        catalog_model(
+            "bigmodel",
+            "glm-4v-flash",
+            "GLM-4V Flash",
+            16_384,
+            1_024,
+            false,
+            true,
+        ),
+        catalog_model(
+            "dashscope",
+            "deepseek-v4-pro",
+            "DeepSeek V4 Pro",
+            1_000_000,
+            384_000,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "deepseek-v4-flash",
+            "DeepSeek V4 Flash",
+            1_000_000,
+            384_000,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "ZHIPU/GLM-5.2",
+            "ZHIPU/GLM-5.2",
+            1_000_000,
+            131_072,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "glm-5.2",
+            "glm-5.2",
+            204_800,
+            128_000,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "glm-5.1",
+            "glm-5.1",
+            202_752,
+            131_072,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "kimi-k2.7-code",
+            "kimi-k2.7-code",
+            262_144,
+            98_304,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope",
+            "kimi-k2.6",
+            "kimi-k2.6",
+            262_144,
+            98_304,
+            true,
+            true,
+        ),
+        catalog_model(
+            "dashscope",
+            "MiniMax/MiniMax-M3",
+            "MiniMax/MiniMax-M3",
+            196_608,
+            32_768,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "MiniMax-M2.5",
+            "MiniMax-M2.5",
+            196_608,
+            32_768,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "mimo-v2.5-pro",
+            "MiMo V2.5 Pro",
+            1_000_000,
+            131_072,
+            true,
+            false,
+        ),
+        catalog_model("dashscope", "glm-5", "glm-5", 202_752, 16_384, true, false),
+        catalog_model(
+            "dashscope",
+            "glm-4.7",
+            "glm-4.7",
+            202_752,
+            16_384,
+            true,
+            false,
+        ),
+        catalog_model(
+            "dashscope",
+            "kimi-k2.5",
+            "kimi-k2.5",
+            262_144,
+            32_768,
+            true,
+            true,
+        ),
+    ]);
+    entries
+}
